organizing types/code
