+++ packages/ui-components/src/components/MetricSeriesChart/index.ts
@@ -1,1 +1,2 @@
 export * from "./MetricSeriesChart";
+export * from "./interfaces";
